Extract bill-to/discount validation to keep persistence under 500 lines
